requirements.txt
dash_bootstrap_components==0.12.2
pandas==1.2.5
plotly==5.1.0
dash_html_components==1.1.3
requests==2.25.1
dash_core_components==1.16.0
dash==1.20.0
python_dateutil==2.8.2
PyYAML==5.4.1
